+++ HISTORY.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.6.2 (unreleased)
+------------------
+- Fix bug that resulted in a ``UnicodeEncodeError`` when tagging text with non-ascii characters.
+- Add ``DecisionTreeClassifier``.
+- Add ``labels()`` and ``train()`` methods to classifiers.
+
 0.6.1 (2013-09-01)
 ------------------
 - Classifiers can be trained and tested on CSV, JSON, or TSV data.